If another hg repository is "locked" or corrupted, or the id can't be made (on the host computer, USB drive, etc.), chorus prevented any send/recieve action. Now it ignores the corrupted repository.
I submit for posterity that the behavior aught to be:

When some repository on the reachable drives is bad (null identifier):

1) Ignore if pushing a new repository (even if it is the same one fixed up)

2) If pushing an expected repository, but it can't be found and there is a bad
one on the target drive, report the fact to the user. (can it be fixed?)

3) Ignore if pushing an expected repository, but it can't be found and there is
a bad one NOT on the target drive, but report that it is not found.

4) If offering a chooser list, add a placeholder item for the bad one.
   If selected, it should give additional info or offer to fix it - if possible.