Allow the client to install a function into the conflict handler to adjust the HtmlDetails of a conflict before it is displayed.
This supports the HTML containing links to objects in FLEx
implemented as database:current. FlexBridge updates these links
to contain the (current) real project name. This strategy prevents
the links breaking if the project is renamed.

I had to provide a rather convoluted access path so FlexBridge
can get access to the MergeConflictEmbeddedMessageContentHandler
to install the function.