Avoid obtaining duplicate projects.
In the USB case, items are disabled if the local projects folder already
has a project with the same folder name as would naturally be created,
OR if it has a project that is already connected to the same repo.

In the ChorusHub case, we disable items where the folder name conflicts,
but only warn the user if we find (after the clone) that another project
uses the same repo ID. There is no way to know this before cloning.

In the Internet case, we already check that the folder the user wants
to use does not exist, and again, we will warn (after the clone) if
another project shares the repo.