Fix a problem with providing context for top-level elements
When we go to report a conflict and context is null, we added a special
case to make one from the conflicting node. This looked for context==null,
which worked for unit tests because the ListenerForUnitTests had a null
context until one was set. It failed in production, because the real
listener has a NullContextDescriptor when it doesn't have a real one.

I changed the unit test listener to be like the real one, and the
critical code to execute also for NullContextDescriptor.

This fixes FieldWorks issue LT-13581.