Enhance notes bar to display notes about multiple targets in multiple files
Multiple targets are supported by the new NotesToRecordMapping function
AdditionalIdGeneratorFunction.

Multiple files are supported by extracting an interface from the
relevant functions of AnnotationRepository and making an additional
implementation of the interface which wraps multiple ones. A new version
of CreateNotesBar (and some of the functions it calls) supports setting
up the model with multiple old-style repos wrapped in the new object.

Also fixed a bug in ChorusSystemTests where the test notes file
was not being created with the right name to match the file annotated.