WS-155 exclude generated files in export folder
There are several "plugins" that generate files in the export
folder that should be excluded from the ProjectFolderConfiguration.

There was code that looked like it was trying to exclude them from
the entire repo, but it seems to only attempt to exclude them from
the root directory.  These patterns were made explicit.