Improve build
Wildcards in item groups below the root element get evaluated when
the file gets read. This change moves the itemgroup under the test
target so that we get all the test files (even the ones that don't
exist at the time the file gets read). Hopefully this will eliminate
the build failures on TC.