Fix ServerSettingsModel.InitFromUri to use UrlDecode
This fixes a bug in the Chorus Internet Settings Dialog which was reported to me by a WeSay user.  The previous code would allow a situation whereby the user has a password with a special character in it like % that when saved into the password field and then the dialog is re-opened, the password has "grown" in length.  This is caused by the password field not being properly url decoded when extracted from the URL.

A regression test is also accompanied.