Present conflict report HTML as UTF-8
In the absence of any <meta> tag specifying an encoding, GeckoFX on
Windows defaults to ISO-8859-1 / CP1252, which leads to garbled text
when a non-Roman word had a conflict. The data is correctly UTF-8
encoded in the Chorus Notes file, so adding an HTML <meta> tag to
specify the encoding is enough to solve the issue.

(cherry picked from commit 17c0cd98dafa8c234fb913a9720e0461601c30c0)