increase MaximumChunkSize from 250KB to 20MB
This will help provide more accurate time estimates as well as reduce number of requests to the server.  This could result in a Send/Receive data transfer speed up.
A "chunk" is a unit of data that is sent across the wire approximately every 5 seconds.  The CalculateChunkSize() adjusts the next chunk size based upon the speed of the previous chunk transmission.  If the previous chunk was sent/received in less than 5 seconds, then the chunk size is increased up to the maximum.  If it took longer than 5 seconds (the target time), then the chunk size is reduced.

The newChunkSize statement was rewritten to do the division first and eliminate 32 bit integer size issues with a large chunkSize.