Always add eol extension
Previously we did this only on Windows, but it makes sense on
Linux as well and helps with line ending normalization when tests
check in a hg repo for unit testing purposes (like LfMerge does).

To normalize line endings in the hg repo, add the extension `eol`
to the `.hgrc` file and tell eol to always normalize line endings:

    [eol]
    only-consistent = False
    [extensions]
    eol =

To have an effect the hg repo also needs a `.hgeol` file:

    [repository]
    native = LF
    [patterns]
    ** = native

If there is no `.hgeol` file in the hg repo the behavior won't
change and no line ending normalization will happen.