Always write timestamps with time zones
In some cases, TimeFormatNoTimeZone was lying, writing out a local time
with a "Z" suffix, which means "This is UTC". This results in loss of
data, since the *actual* time of the original event (conflict, message,
etc.) can no longer be retrieved. If it's important to write out local
time rather than UTC in a timestamp, then that timestamp should carry
the actual time zone it was created in, rather than a false Z suffix.