Correct project TC bt to master
Set TC build_type to the TC job building on master, to match this
branch. Then we will download the correct dependencies and build
according to APIs that clients are expecting us to be using.

For example, recognize that SIL.UsbDrive types were moved from
SIL.Core.dll to SIL.Core.Deskotp.dll.

This mimics what was done in buildupdate.win.sh in recent commit
13c964894.