Adjust for new L10NSharp version
L10NSharp now supports selecting the translation memory in the API
rather than by having two different branches.

This change adds a new overload of `ChorusSystem.SetUpLocalization`
that allows to select the translation memory, and changes the
existing method to use TMX files, which is what we had previously.
This is necessary to retain previous functionality - previously
you could exchange the L10NSharp.dll to select the translation
memory. Now this can be done by passing an additional parameter
to `SetupLocalization`.