Fix bitness of executables
We have to set PlatformTarget explicitly, otherwise msbuild will make
the executables x86 because icu4c contains platform dependent native
binaries. However, our icu.net library expects to find the binaries
for both x86 and x64, so using AnyCPU for the executable is
acceptable. For similar reasons we have to set RuntimeIdentifiers
which is needed when building with .NET 5 on AppVeyor - without it
we get an error that there is no target for 'net461/win7-x86'.