Speed up Send/Receive operations significantly
The Send/Receive code does a lot of `hg log -rREV --template "{node}"`
calls in order to learn the long hash of a revision number. This happens
very often during a Send/Receive process: the Chorus logs are full of
`hg log -r0` operations. Thing is, every single call to `hg log` incurs
a cost as a Python process has to be spun up every time this happens.
And in any given repository, a specific revision number wil always refer
to the same revision with the same long hash, so each long hash only has
to be calculated *once* for any given revision number. So we'll cache
the output of `log -rREV` operations and return the cached output if the
same operation runs a second time. (Other operations like `hg parents`
aren't safe to cache, as their result will change during a Send/Receive
operation as new commits are added to the repo. Only `log -rREV` calls
are 100% safe to cache).