Make LGTM checker happy
Elsewhere the `command` parameter is used without checking for nullity,
so LGTM.com considers my null check to add a new possible problem. Since
this is an internal method and we're pretty much never going to pass a
null command string in here, we can remove that null check and be okay.