Don't cache all-zero results from hg
An all-zero result from hg is highly likely to be the repo identifier of
an empty repository, which is going to change in the near future. So
we'll skip caching anything that starts with twelve zeroes in a row.
This will occasionally have a false positive, but that false positive
won't have any negative impact (a minor slowdown).