Speed up Send/Receive operations significantly (#258)
The Send/Receive code does a lot of `hg log -rREV --template "{node}"`
calls in order to learn the long hash of a revision number. This happens
very often during a Send/Receive process: the Chorus logs are full of
`hg log -r0` operations. Thing is, every single call to `hg log` incurs
a cost as a Python process has to be spun up every time this happens.
And in any given repository, a specific revision number wil always refer
to the same revision with the same long hash, so each long hash only has
to be calculated *once* for any given revision number. So we'll cache
the output of `log -rREV` operations and return the cached output if the
same operation runs a second time. (Other operations like `hg parents`
aren't safe to cache, as their result will change during a Send/Receive
operation as new commits are added to the repo. Only `log -rREV` calls
are 100% safe to cache).

* Make ExecuteErrorsOk an instance method

Make ExecuteErrorsOk an instance method since there's no reason it needs
to be static (no code that uses the Chorus library can be calling it
since it's private, so this won't affect any other code). This also
allows simplifying its type signature.

* Don't cache all-zero results from hg

An all-zero result from hg is highly likely to be the repo identifier of
an empty repository, which is going to change in the near future. So
we'll skip caching anything that starts with twelve zeroes in a row.
This will occasionally have a false positive, but that false positive
won't have any negative impact (a minor slowdown).

* Do not cache negative revision numbers

Negative revision numbers mean "counting backwards from tip", i.e.
running `hg log -r-1 --template "{node}" gets the long hash of the tip
revision. We therefore should not cache negative revision numbers as
those are subject to change.