LT-20999: Move the location of the local Chorus cache
Move the location of the local Chorus cache from AppData to the
specific FLEX or WeSay project.  The problem we were seeing was when
the cache in AppData was being used by two applications; in this
case FLEX and WeSay. (Note that the same problem would exist if two
local WeSay projects or two local FLEX project were using the same
Chorus Project ID.)
When the cache was more up to date than the local FLEX or WeSay
project we would see the error described in the defect. To prevent
this problem we now store the cache in the local FLEX or WeSay
project folder, along with the other project information. This
results in one cache per application/project, so the cache should
never have more up to date chorus data than the local FLEX or
WeSay project.