increase audio and picture file size limit to 10MB
This is in response to users who report a need to add audio and pictures to their project over 1MB.  The 1MB is reported to be too small.

Because we don't know practically how this will impact the majority of users, I am happy to introduce this change and evaluate the impact (plus/minus) on Language Forge and FLEx users who use Language Depot to collaborate.