Add public StartClone() method (#325)
This is necessary so that Flex Bridge can supply required cloning information and start a repo download without direct/manual user input. This order will come from Fieldworks when launched from CMD or protocol handler with the proper arguments.

Required for https://github.com/sillsdev/languageforge-lexbox/issues/242

+semver:major

* Centralise CloneResult generation and simplify DoClone

* Revert user/pass after API clone

We want to give callers the option to revert, e.g. if they know the user/pass is JWT, but default it to saving them.

Co-Authored-By: Hasso <hasso_pape@sil.org>