Add hg username to four unit tests
A behavior change between Mercurial 3.3 and 6.5.1 means that these four
tests now fail due to Mercurial complaining that the username isn't set.
Since the absence of a username has nothing to do with the matter under
test, we should just change how these four tests are set up to include a
username so that Mercurial is satisfied.