Restore flaky test to original timing and ignore it
Previous PR switched this test to run for 25s instead of 10s. This was
still not enough to make it work on the GitHub Actions runner, so we
will just start ignoring it from now on. It's unreliable enough that it
really shouldn't be part of a unit test suite anyway.