Temporarily disable tests in CI
The tests take two hours to run now that we have net461, net6.0, and
net8.0 target frameworks. They're all passing except for one that's
known to be flaky and timing-dependant, so it's safe to disable them
while we work on fixing the part of the build that comes after the
tests. Then once that's working we'll revert this commit.