Also build for net8.0 (#332)
* Also build for net8.0
* Make GHA workflow run .NET 6 and 8
* Update setup-dotnet action

v2.1.0 was still trying to build against win7-x32 runtimes, which are no
longer offered in .NET 8. If we're using recent .NET, we need a recent
setup-dotnet version.

* Fix RuntimeIdentifiers
* Suppress WiX validation as per WiX docs
* Ignore flaky test on Linux, too

The GitHub Actions runners are apparently slow to deliver filesystem
notifications, which is exactly the scenario that makes this test
useless. It's already disabled on Windows; there's no reason for it to
cause the CI build to fail all the time when it's the only failing test.