Replace Platform.IsMono with actual OS checks
Some cases of Platform.IsMono are workarounds for bugs in older versions
of mono. Those can probably be removed, but for safety's sake, we'll
leave them alone for now.

Other cases are clearly trying to check the platform, to run Linux-only
or Windows-only code. Those cases should be replaced by Platform.IsUnix
or Platform.IsWindows as appropriate, because nearly everywhere that
Chorus is being run on Linux these days, it's running under dotnet
rather than mono, and the Platform.IsMono check returns false.