Run tests in parallel, one per framework (#354)
* Run tests in parallel, one per framework

Now that tests run on three frameworks (net461, net6.0, net8.0), we can
save a lot of build time by running each framework's tests in parallel.

* Ensure parallel test runs don't cancel each other

* Don't specify framework for `dotnet build` step

* Attempt to work around `dotnet test` bug

If running with `--no-build`, then `dotnet test` complains about DLLs
being invalid arguments, resulting in an exit code of 1 even if all
tests passed. This then makes GHA think the build was a failure. The
discussion in the dotnet SDK issue suggests removing `--no-build` as
one possible workaround for this issue.

* Try doing restore-build-test in one go on Linux

Otherwise we get errors about the dotnet restore not being done for the
right framework.

* Try restoring for specific target framework

* Try specifying platform on Linux

Also, forgot to specify target framework on Windows; no wonder it was
taking so much longer!

* Try not specifying runtime for Linux

* ChorusMerge tests should run on all frameworks

Since ChorusMerge runs on net461;net6.0;net8.0, the ChorusMerge tests
should be run on all the same frameworks.

* Explicitly specify test project per step

Also only run Chorus.Tests on net461, since it's not defined for net6.0
and net8.0.

* Correct path to LibChorus tests

* lowered NUnit3TestAdapter version to a version which supports net461

* run Chorus hub tests in CI

* Checkout sources before building installer

Now that the installer job is not running in the same job matrix as the
tests, we need to re-run the dotnet restore and build steps again.

* Remove net6.0 targets

Everything that uses Chorus now runs on net8.0 or higher.

---------

Co-authored-by: Kevin Hahn <kevin_hahn@sil.org>